+++ payments/src/sdk/models/shared/frequencyupdate.ts
@@ -8,6 +8,7 @@ import { Expose } from "class-transformer";
 export enum FrequencyUpdateInterval {
     Week = "Week",
     Month = "Month",
+    LessThanNilGreaterThan = "<nil>",
 }
 
 export class FrequencyUpdate extends SpeakeasyBase {