+++ payments/src/sdk/models/shared/invoiceupdaterequest.ts
@@ -18,6 +18,7 @@ export enum InvoiceUpdateRequestAcceptedPaymentMethods {
     Credit = "Credit",
     Ach = "ACH",
     Manual = "Manual",
+    LessThanNilGreaterThan = "<nil>",
 }
 
 export class InvoiceUpdateRequestClient extends SpeakeasyBase {}
@@ -32,6 +33,7 @@ export enum InvoiceUpdateRequestStatus {
     Pending = "Pending",
     PaymentInTransit = "PaymentInTransit",
     Paid = "Paid",
+    LessThanNilGreaterThan = "<nil>",
 }
 
 export class InvoiceUpdateRequest extends SpeakeasyBase {