+++ payments/RELEASES.md
@@ -9,3 +9,13 @@ Based on:
 - [typescript v2.0.0] payments
 ### Releases
 - [NPM v2.0.0] https://www.npmjs.com/package/@wingspan/payments/v/2.0.0 - payments
+
+## 2023-10-31 13:48:10
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.wingspan.app/openapi/646bee8476ac6400f0ed8faa
+- Speakeasy CLI 1.109.0 (2.173.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v1.1.0] payments
+### Releases
+- [NPM v1.1.0] https://www.npmjs.com/package/@wingspan/payments/v/1.1.0 - payments